Minor changes in configuration of the project, cleaned up code
