+++ app/src/main/java/com/example/fbuteamproject/activities/ARActivity.java
@@ -2,28 +2,39 @@ package com.example.fbuteamproject.activities;
 
 import android.Manifest;
 import android.app.Activity;
-import android.content.pm.PackageManager;
+import android.graphics.SurfaceTexture;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.RequiresApi;
 import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.GestureDetector;
 import android.view.MotionEvent;
 import android.view.View;
+import android.widget.EditText;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 
 import com.example.fbuteamproject.R;
-import com.example.fbuteamproject.fragments.NoteContentFragment;
-import com.example.fbuteamproject.interfaces.PassNoteToActivityListener;
-import com.example.fbuteamproject.models.Note;
+import com.example.fbuteamproject.layouts.ARComponentsShell;
 import com.example.fbuteamproject.models.Planet;
+import com.example.fbuteamproject.utils.DemoUtils;
+import com.google.android.exoplayer2.DefaultLoadControl;
+import com.google.android.exoplayer2.DefaultRenderersFactory;
+import com.google.android.exoplayer2.ExoPlayerFactory;
+import com.google.android.exoplayer2.SimpleExoPlayer;
+import com.google.android.exoplayer2.source.ExtractorMediaSource;
+import com.google.android.exoplayer2.source.MediaSource;
+import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
+import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory;
+import com.google.android.exoplayer2.util.Util;
+import com.google.android.exoplayer2.video.VideoListener;
 import com.google.ar.core.Anchor;
-import com.google.ar.core.ArCoreApk;
-import com.google.ar.core.Config;
 import com.google.ar.core.Frame;
 import com.google.ar.core.HitResult;
 import com.google.ar.core.Plane;
@@ -37,67 +48,109 @@ import com.google.ar.sceneform.ArSceneView;
 import com.google.ar.sceneform.HitTestResult;
 import com.google.ar.sceneform.Node;
 import com.google.ar.sceneform.math.Vector3;
+import com.google.ar.sceneform.rendering.Color;
+import com.google.ar.sceneform.rendering.ExternalTexture;
 import com.google.ar.sceneform.rendering.ModelRenderable;
 import com.google.ar.sceneform.rendering.Renderable;
 import com.google.ar.sceneform.rendering.ViewRenderable;
 
-import org.parceler.Parcels;
-
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
-public class ARActivity extends AppCompatActivity implements PassNoteToActivityListener {
+import static com.example.fbuteamproject.utils.DemoUtils.checkIsSupportedDeviceOrFinish;
 
-    //TODO - Bring this back if needed
-//    private FragmentManager fragmentManager;
+/* ARActivity that populates the sceneview with model renderables and view renderables
+for notes, photos and videos upon clicking on the screen.
+*/
+public class ARActivity extends AppCompatActivity {
 
-    private final String APP_TAG = "NoteActivity";
-    private static final int RC_PERMISSIONS = 0x123;
+    private static final String TAG = ARActivity.class.getSimpleName();
+    private static final int RC_PERMISSIONS =0x123;
+
     private boolean installRequested;
+    
+    @Nullable
+    private ModelRenderable videoRenderable;
+    private ModelRenderable venusRenderable;
+    private ModelRenderable jupiterRenderable;
 
-
-    private ArSceneView arSceneView;
-    private GestureDetector gestureDetector;
-    private Snackbar loadingMessageSnackbar = null;
-
-    CompletableFuture<ModelRenderable> earthStage;
-    CompletableFuture<ModelRenderable> marsStage;
-    CompletableFuture<ModelRenderable> neptuneStage;
-    CompletableFuture<ViewRenderable> planetTitleStage;
-    CompletableFuture<ViewRenderable> planetContentsStage;
-
-
-
-    private ModelRenderable earthRenderable;
-    private ModelRenderable marsRenderable;
-    private ModelRenderable neptuneRenderable;
+    // button renderables
+    private ViewRenderable buttonPauseRenderable;
+    private ViewRenderable buttonResumeRenderable;
+    private ViewRenderable buttonStopRenderable;
 
     private ViewRenderable planetTitlesRenderable;
     private ViewRenderable planetContentsRenderable;
 
+    //Initialize ExoPlayer variables
+    private SimpleExoPlayer player;
+    private boolean playWhenReady;
+    private int currentWindow = 0;
+    private long playbackPosition = 0;
+
+    private GestureDetector gestureDetector;
+
+    private ArSceneView arSceneView;
+
+    private Snackbar loadingMessageSnackbar;
+
     // True once scene is loaded
-    private boolean hasFinishedLoading = false;
+    private boolean hasFinishedLoading;
 
     // True once the scene has been placed.
-    private boolean hasPlacedComponents = false;
+    private boolean hasPlacedComponents;
+
+    // The color to filter out of the video.
+    private static final Color CHROMA_KEY_COLOR = new Color(0.1843f, 1.0f, 0.098f);
+    // Controls the height of the video in world space.
+    private static final float VIDEO_HEIGHT_METERS = 0.7f;
+
+    //Completable Futures for model renderables
+    CompletableFuture<ModelRenderable> venusStage;
+    CompletableFuture<ModelRenderable> jupiterStage;
+    CompletableFuture<ModelRenderable> videoStage;
+    CompletableFuture<ViewRenderable> buttonPauseStage;
+    CompletableFuture<ViewRenderable> buttonResumeStage;
+    CompletableFuture<ViewRenderable> buttonStopStage;
+
+    // viewrenderables for photos
+    private ViewRenderable photoRenderable1;
+    private ViewRenderable photoRenderable2;
+    private ViewRenderable photoRenderable3;
+    private ViewRenderable photoRenderable4;
 
 
-//    private ArrayList<Note> notesTest;
-//    private NoteTitleAdapter noteTitleAdapter;
+    CompletableFuture<ViewRenderable> photoStage1;
+    CompletableFuture<ViewRenderable> photoStage2;
+    CompletableFuture<ViewRenderable> photoStage3;
+    CompletableFuture<ViewRenderable> photoStage4;
 
+    CompletableFuture<ViewRenderable> planetTitleStage;
+    CompletableFuture<ViewRenderable> planetContentsStage;
 
     @RequiresApi(api = Build.VERSION_CODES.N)
     @Override
+    @SuppressWarnings({"AndroidApiChecker", "FutureReturnValueIgnored"})
+    // CompletableFuture requires api level 24
+    // FutureReturnValueIgnored is not valid
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        //TODO - Make sure to change this when finished
-        setContentView(R.layout.activity_ar);
+        if (!checkIsSupportedDeviceOrFinish(this)) {
+            return;
+        }
 
+        //set the content and the layout
+        setContentView(R.layout.activity_videos);
+
+        //find the sceneview
         arSceneView = findViewById(R.id.ar_scene_view);
 
-        buildModelRenderableFutures();
-        buildViewRenderableFutures();
+        buildPlanetRenderables();
+
+        buildVideoRenderable();
+
+        buildViewRenderables();
 
         setupRenderables();
 
@@ -105,24 +158,13 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
         setupTouchListener();
         setupOnUpdateListener();
 
+        // Lastly request CAMERA permission which is required by ARCore.
         requestCameraPermission(this);
-
-
-        //TODO = Bring this back if needed
-//        fragmentManager = getSupportFragmentManager();
-//
-//        NoteTitleFragment initFragment = new NoteTitleFragment();
-//
-//        initFragment.setPassNoteListener(this);
-//
-//        fragmentManager
-//                .beginTransaction()
-//                .add(R.id.fragmentContainer, initFragment)
-//                .commit();
-
     }
 
     private void setupOnUpdateListener() {
+        // Set an update listener on the Scene that will hide the loading message once a Plane is
+        // detected.
         arSceneView
                 .getScene()
                 .addOnUpdateListener(
@@ -149,6 +191,7 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
     }
 
     private void setupTouchListener() {
+        // Set a touch listener on the Scene to listen for taps.
         arSceneView
                 .getScene()
                 .setOnTouchListener(
@@ -156,25 +199,23 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
                             // If the solar system hasn't been placed yet, detect a tap and then check to see if
                             // the tap occurred on an ARCore plane to place the solar system.
                             if (!hasPlacedComponents) {
-                                Log.d(APP_TAG, "Wanting to place Fragments for first time");
                                 return gestureDetector.onTouchEvent(event);
                             }
 
                             // Otherwise return false so that the touch event can propagate to the scene.
-                            Log.d(APP_TAG, "The Fragments are already out there, so we wanna" +
-                                    " do stuff in the scene");
                             return false;
                         });
     }
 
     private void setupGestureDetector() {
+        //Gesture detector for tap events
         gestureDetector =
                 new GestureDetector(
                         this,
                         new GestureDetector.SimpleOnGestureListener() {
+                            @RequiresApi(api = Build.VERSION_CODES.N)
                             @Override
                             public boolean onSingleTapUp(MotionEvent e) {
-                                Log.d(APP_TAG, "Just released my finger after having tapped");
                                 onSingleTap(e);
                                 return true;
                             }
@@ -186,43 +227,94 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
                         });
     }
 
+    @RequiresApi(api = Build.VERSION_CODES.N)
     private void setupRenderables() {
         CompletableFuture.allOf(
-                    earthStage,
-                    marsStage,
-                    neptuneStage,
-                    planetTitleStage,
-                    planetContentsStage)
+                videoStage,
+                venusStage,
+                jupiterStage, photoStage1, photoStage2, photoStage3, photoStage4,
+                buttonPauseStage,
+                buttonResumeStage,
+                buttonStopStage,
+                planetTitleStage,
+                planetContentsStage)
                 .handle(
                         (notUsed, throwable) -> {
-                            // When you build a Renderable, Sceneform loads its resources in the background while
-                            // returning a CompletableFuture. Call handle(), thenAccept(), or check isDone()
-                            // before calling get().
-
                             if (throwable != null) {
-                                Log.e(APP_TAG, "Throwable found to be non-null in handle method");
+                                DemoUtils.displayError(this, "Unable to load renderable", throwable);
                                 return null;
                             }
-
                             try {
-                                earthRenderable = earthStage.get();
-                                marsRenderable = marsStage.get();
-                                neptuneRenderable = neptuneStage.get();
+                                videoRenderable = videoStage.get();
+                                venusRenderable = venusStage.get();
+                                jupiterRenderable = jupiterStage.get();
+                                buttonPauseRenderable = buttonPauseStage.get();
+                                buttonResumeRenderable = buttonResumeStage.get();
+                                buttonStopRenderable = buttonStopStage.get();
+
+                                photoRenderable1 = photoStage1.get();
+                                photoRenderable2 = photoStage2.get();
+                                photoRenderable3 = photoStage3.get();
+                                photoRenderable4 = photoStage4.get();
+
                                 planetTitlesRenderable = planetTitleStage.get();
                                 planetContentsRenderable = planetContentsStage.get();
 
                                 // Everything finished loading successfully.
                                 hasFinishedLoading = true;
-
                             } catch (InterruptedException | ExecutionException ex) {
-                                Log.e(APP_TAG, "Unable to load Renderables (IN CATCH STATEMENT)");
+                                DemoUtils.displayError(this, "Unable to load renderable", ex);
                             }
-
                             return null;
                         });
     }
 
-    private void buildViewRenderableFutures() {
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private void buildPlanetRenderables() {
+        venusStage =
+                ModelRenderable
+                .builder()
+                .setSource(this, Uri.parse("Venus_1241.sfb"))
+                .build();
+        jupiterStage =
+                ModelRenderable
+                        .builder()
+                        .setSource(this, Uri.parse("model.sfb"))
+                        .build();
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private void buildVideoRenderable() {
+        videoStage =
+                ModelRenderable
+                        .builder()
+                        .setSource(this, R.raw.video_screen)
+                        .build();
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private void buildViewRenderables() {
+        buttonPauseStage =
+                ViewRenderable
+                        .builder()
+                        .setView(this, R.layout.buttonpause)
+                        .build();
+        buttonResumeStage =
+                ViewRenderable
+                        .builder()
+                        .setView(this, R.layout.buttonresume)
+                        .build();
+        buttonStopStage =
+                ViewRenderable
+                        .builder()
+                        .setView(this, R.layout.buttonstop)
+                        .build();
+
+        photoStage1 = ViewRenderable.builder().setView(this, R.layout.test_ar1).build();
+        photoStage2 = ViewRenderable.builder().setView(this, R.layout.test_ar2).build();
+        photoStage3 = ViewRenderable.builder().setView(this, R.layout.test_ar1).build();
+        photoStage4 = ViewRenderable.builder().setView(this, R.layout.test_ar1).build();
+
         planetTitleStage =
                 ViewRenderable
                         .builder()
@@ -233,23 +325,6 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
                 ViewRenderable
                         .builder()
                         .setView(this, R.layout.component_planet_contents)
-                        .build();
-    }
-
-    private void buildModelRenderableFutures() {
-        earthStage =
-                ModelRenderable
-                        .builder()
-                        .setSource(this, Uri.parse("CHAHIN_EARTH.sfb") )
-                        .build();
-        marsStage =
-                ModelRenderable
-                        .builder()
-                        .setSource(this, Uri.parse("1239 Mars.sfb") )
-                        .build();
-        neptuneStage = ModelRenderable
-                        .builder()
-                        .setSource(this, Uri.parse("Neptune.sfb") )
                         .build();
     }
 
@@ -264,31 +339,31 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
             // If the session wasn't created yet, don't resume rendering.
             // This can happen if ARCore needs to be updated or permissions are not granted yet.
             try {
-                Session session = createArSession(this, installRequested);
+                Session session = DemoUtils.createArSession(this, installRequested);
                 if (session == null) {
-                    installRequested = hasCameraPermission(this);
+                    installRequested = DemoUtils.hasCameraPermission(this);
                     return;
                 } else {
                     arSceneView.setupSession(session);
+
+
                 }
             } catch (UnavailableException e) {
-                Log.e(APP_TAG, "Some stuff went down...\n" + e.getLocalizedMessage() );
+                DemoUtils.handleSessionException(this, e);
             }
         }
 
         try {
             arSceneView.resume();
         } catch (CameraNotAvailableException ex) {
-            Log.e(APP_TAG, "Unable to get camera...");
+            DemoUtils.displayError(this, "Unable to get camera", ex);
             finish();
             return;
         }
 
-
         if (arSceneView.getSession() != null) {
             showLoadingMessage();
         }
-
     }
 
     @Override
@@ -297,121 +372,269 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
         if (arSceneView != null) {
             arSceneView.pause();
         }
+        if (player != null) {
+            releasePlayer();
+        }
     }
 
     @Override
     public void onDestroy() {
         super.onDestroy();
-        if (arSceneView != null) {
-            arSceneView.destroy();
+
+        if (player != null) {
+            releasePlayer();
         }
     }
 
-
-
+    @RequiresApi(api = Build.VERSION_CODES.N)
     private void onSingleTap(MotionEvent tap) {
         if (!hasFinishedLoading) {
             // We can't do anything yet.
-            Log.d(APP_TAG, "Not done loading Scene yet");
             return;
         }
-
-        Log.d(APP_TAG, "In onSingleTap and have finished loading the Scene");
 
         Frame frame = arSceneView.getArFrame();
         if (frame != null) {
             if (!hasPlacedComponents && tryPlaceComponents(tap, frame)) {
-                hasPlacedComponents = true;
+                hasPlacedComponents= true;
             }
         }
     }
 
-
-    private boolean tryPlaceComponents(MotionEvent tap, Frame frame) {
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private boolean tryPlaceComponents (MotionEvent tap, Frame frame) {
         if (tap != null && frame.getCamera().getTrackingState() == TrackingState.TRACKING) {
             for (HitResult hit : frame.hitTest(tap)) {
                 Trackable trackable = hit.getTrackable();
                 if (trackable instanceof Plane && ((Plane) trackable).isPoseInPolygon(hit.getHitPose())) {
-                    Log.d(APP_TAG, "Creating anchor to try and place Fragments");
                     setupAnchor(hit);
                     return true;
                 }
-
             }
-
         }
-
         return false;
     }
 
+    @RequiresApi(api = Build.VERSION_CODES.N)
     private void setupAnchor(HitResult hit) {
+        // Create the Anchor.
         Anchor anchor = hit.createAnchor();
         AnchorNode anchorNode = new AnchorNode(anchor);
         anchorNode.setParent(arSceneView.getScene());
-        Node fragments = createComponents();
-        anchorNode.addChild(fragments);
+        Node components = createComponents();
+        components.setParent(anchorNode);
     }
 
-
+    @RequiresApi(api = Build.VERSION_CODES.N)
     private Node createComponents() {
 
-        Log.d(APP_TAG, "In Component creation method");
+        Planet venusVisual = new Planet("Venus", "Venus is a goddess", getString(R.string.venus_res), this );
+        venusVisual.setRenderable(venusRenderable);
 
-        Node base = new Node();
+        Planet jupiterVisual = new Planet("Jupiter", "Jupiter is a god", getString(R.string.jupiter_res), this);
+        jupiterVisual.setRenderable(jupiterRenderable);
 
-//        Node noteTitles = new Node();
-//        noteTitles.setParent(base);
+        Node photo1 = new Node();
+        photo1.setRenderable(photoRenderable1);
 
-        Node planetTitles = new Node();
-        setupNode(planetTitles, base, planetTitlesRenderable, new Vector3(-0.5f, 0.5f, 0.0f), new Vector3(0.5f, 0.35f, 0.5f) );
+        Node photo2 = new Node();
+        photo2.setRenderable(photoRenderable1);
+
+        Node photo3 = new Node();
+        photo3.setRenderable(photoRenderable1);
+
+        Node photo4 = new Node();
+        photo4.setRenderable(photoRenderable2);
+
+        Node photo5 = new Node();
+        photo5.setRenderable(photoRenderable2);
+
+        Node photo6 = new Node();
+        photo6.setRenderable(photoRenderable2);
+
+        Node videoNode = new Node();
 
         Node planetContents = new Node();
-        setupNode(planetContents, base, planetContentsRenderable, new Vector3(0.5f, 0.5f, 0.0f), new Vector3(0.5f, 0.35f, 0.5f) );
+        planetContents.setRenderable(planetContentsRenderable);
 
-//        Planet earthVisual = new Planet("Earth", "The grandest of places! I hear all the life is here ;)");
-//        setupNode(earthVisual, base, earthRenderable, new Vector3(-0.5f, 1.5f, 0.0f), new Vector3(0.2f, 0.2f, 0.2f) );
-//
-//        Planet marsVisual = new Planet("Mars","Pretty nifty place. Elon likes it a lot, I'm pretty sure :0");
-//        setupNode(marsVisual, base, marsRenderable, new Vector3(0.0f, 1.5f, 0.0f), new Vector3(0.2f, 0.2f, 0.2f) );
-//
-//        Planet neptuneVisual = new Planet("Neptune","King of the Sea?? More like last planet in our Solar System LMAO XD");
-//        setupNode(neptuneVisual, base, neptuneRenderable, new Vector3(0.5f, 1.5f, 0.0f), new Vector3(0.2f, 0.2f, 0.2f) );
-
+        //Organizes all the components relative to each other
+        Node base = new ARComponentsShell(venusVisual, jupiterVisual, photo1,
+                photo2, photo3, photo4, photo5, photo6, videoNode, planetContents);
 
         View planetTitleView = planetTitlesRenderable.getView();
         View planetContentView = planetContentsRenderable.getView();
 
-//        setupPlanetTapListeners(earthVisual, marsVisual, neptuneVisual, planetTitleView, planetContentView);
+
+        setupPlanetTapListenerVideo(venusVisual, jupiterVisual, base, planetTitleView, planetContentView);
 
         return base;
     }
 
-    private void setupPlanetTapListeners(Planet earthVisual, Planet marsVisual, Planet neptuneVisual, View planetTitleView, View planetContentView) {
-        earthVisual.setOnTapListener((hitTestResult, motionEvent) -> {
-//            changePlanetScreenText(planetTitleView, planetContentView, earthVisual);
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private void setupPlanetTapListenerVideo(Planet venusVisual, Planet jupiterVisual, Node baseNode, View planetTitleView, View planetContentView) {
+
+        // Create an ExternalTexture for displaying the contents of the video.
+        ExternalTexture texture = new ExternalTexture();
+
+        venusVisual.setOnTapListener((hitTestResult, motionEvent) -> {
+
+            playVideo(venusVisual, baseNode, texture);
+            changePlanetScreenText(planetTitleView, planetContentView, venusVisual);
         });
 
-        marsVisual.setOnTapListener((hitTestResult, motionEvent) -> {
-//            changePlanetScreenText(planetTitleView, planetContentView, marsVisual);
+        jupiterVisual.setOnTapListener((hitTestResult, motionEvent) -> {
+
+            playVideo(jupiterVisual, baseNode, texture);
+            changePlanetScreenText(planetTitleView, planetContentView, jupiterVisual);
+
         });
 
-        neptuneVisual.setOnTapListener((hitTestResult, motionEvent) -> {
-//            changePlanetScreenText(planetTitleView, planetContentView, neptuneVisual);
-        });
     }
 
     private void changePlanetScreenText(View nameView, View contentView, Planet currPlanet){
-//        ((TextView) nameView.findViewById(R.id.tvTitle) ).setText(currPlanet.getPlanetName() );
-//        ((TextView) contentView.findViewById(R.id.tvContents) ).setText(currPlanet.getPlanetNotes() );
+        ((TextView) nameView).setText(currPlanet.getPlanetName() );
+        ((EditText) contentView).setText(currPlanet.getPlanetNotes() );
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private void playVideo(Planet planetVisual, Node baseNode, ExternalTexture texture) {
+
+        stopPlaying();
+
+        setupExoPlayer(texture, planetVisual.getPlanetVideoResID());
+
+        Node video = getVideoNode(baseNode);
+
+        setVideoTexture(texture);
+
+        startExoPlayer(texture, video);
+    }
+
+    private void setupExoPlayer(ExternalTexture texture, String videoResID) {
+
+        player = ExoPlayerFactory.newSimpleInstance(
+                new DefaultRenderersFactory(this),
+                new DefaultTrackSelector(), new DefaultLoadControl());
+
+        player.setVideoSurface(texture.getSurface());
+
+        player.setPlayWhenReady(playWhenReady);
+        player.seekTo(currentWindow, playbackPosition);
+
+        Uri uri = Uri.parse(videoResID);
+        MediaSource mediaSource = buildMediaSource(uri);
+        player.prepare(mediaSource, true, false);
+
+        player.addVideoListener(new VideoListener() {
+            @Override
+            public void onVideoSizeChanged(int width, int height, int unappliedRotationDegrees, float pixelWidthHeightRatio) {
+                Log.d(TAG, "play video listener video size changed");
+            }
+            @Override
+            public void onRenderedFirstFrame() {
+                Log.d(TAG,"play video listener render first frame");
+            }
+        });
+    }
+
+    private MediaSource buildMediaSource(Uri uri) {
+        return new ExtractorMediaSource.Factory(
+                new DefaultHttpDataSourceFactory(Util.getUserAgent(this, "minerva"))).
+                createMediaSource(uri);
+    }
+
+    private Node getVideoNode(Node baseNode) {
+        Node video = new Node();
+
+        setupNode(video, baseNode, videoRenderable, new Vector3(0.0f, 0.5f, 0.0f), new Vector3(
+                VIDEO_HEIGHT_METERS * 2, VIDEO_HEIGHT_METERS, 1.0f));
+        return video;
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    private void setVideoTexture(ExternalTexture texture) {
+            videoRenderable.getMaterial().setExternalTexture("videoTexture", texture);
+            videoRenderable.getMaterial().setFloat4("keyColor", CHROMA_KEY_COLOR);
     }
 
 
+    private void startExoPlayer(ExternalTexture texture, Node video) {
+        if (!player.getPlayWhenReady()) {
+
+            player.setPlayWhenReady(true);
+
+            texture.getSurfaceTexture().setOnFrameAvailableListener(
+                        (SurfaceTexture surfaceTexture) -> {
+                            video.setRenderable(videoRenderable);
+                            texture.getSurfaceTexture().setOnFrameAvailableListener(null);
+                        });
+        } else {
+        video.setRenderable(videoRenderable);
+        }
+    }
 
     private void setupNode(Node currNode, Node baseNode, Renderable renderable, Vector3 localPos, Vector3 localScale){
         currNode.setParent(baseNode);
         currNode.setRenderable(renderable);
         currNode.setLocalPosition(localPos);
         currNode.setLocalScale(localScale);
+    }
+
+    public void pause(View v) {
+        if (player == null) {
+            Toast.makeText(this, "Video has been stopped", Toast.LENGTH_LONG).show();
+            return;
+        }
+        if (!player.getPlayWhenReady()) {
+            Toast.makeText(this, "Video has been paused", Toast.LENGTH_LONG).show();
+            return;
+        }
+
+        player.setPlayWhenReady(!player.getPlayWhenReady());
+    }
+
+    public void resume(View v) {
+
+        if (player == null) {
+            Toast.makeText(this,"Video is playing" , Toast.LENGTH_LONG).show();
+            return;
+        }
+        if (player.getPlayWhenReady()) {
+            Toast.makeText(this, "Video has been paused", Toast.LENGTH_LONG).show();
+            return;
+        }
+
+        player.setPlayWhenReady(true);
+    }
+
+    private void stopPlaying() {
+        releasePlayer();
+    }
+
+    private void releasePlayer() {
+        if (player == null) {
+            return;
+        }
+
+        playbackPosition = player.getCurrentPosition();
+        currentWindow = player.getCurrentWindowIndex();
+        playWhenReady = player.getPlayWhenReady();
+        player.release();
+        player = null;
+
+    }
+
+    public void stop(View v) {
+        stopPlaying();
+    }
+
+    private void hideLoadingMessage() {
+        if (loadingMessageSnackbar == null) {
+            return;
+        }
+
+        loadingMessageSnackbar.dismiss();
+        loadingMessageSnackbar = null;
     }
 
     private void showLoadingMessage() {
@@ -428,74 +651,10 @@ public class ARActivity extends AppCompatActivity implements PassNoteToActivityL
         loadingMessageSnackbar.show();
     }
 
-    private void hideLoadingMessage() {
-        if (loadingMessageSnackbar == null) {
-            return;
-        }
-
-        loadingMessageSnackbar.dismiss();
-        loadingMessageSnackbar = null;
-    }
-
-
-    private Session createArSession(Activity activity, boolean installRequested)
-            throws UnavailableException {
-        Session session = null;
-        // if we have the camera permission, create the session
-        if (hasCameraPermission(activity)) {
-            switch (ArCoreApk.getInstance().requestInstall(activity, !installRequested)) {
-                case INSTALL_REQUESTED:
-                    return null;
-                case INSTALLED:
-                    break;
-            }
-            session = new Session(activity);
-            // IMPORTANT!!!  ArSceneView requires the `LATEST_CAMERA_IMAGE` non-blocking update mode.
-            Config config = new Config(session);
-            config.setUpdateMode(Config.UpdateMode.LATEST_CAMERA_IMAGE);
-            session.configure(config);
-        }
-        return session;
-    }
-
-    private boolean hasCameraPermission(Activity activity) {
-        return ContextCompat.checkSelfPermission(activity, Manifest.permission.CAMERA)
-                == PackageManager.PERMISSION_GRANTED;
-    }
 
     private void requestCameraPermission(Activity activity) {
         ActivityCompat.requestPermissions(
                 activity, new String[] {Manifest.permission.CAMERA}, ARActivity.RC_PERMISSIONS);
     }
-
-
-
-
-    //TODO - Bring this back if needed
-    @Override
-    public void returnSelectedNote(Note selectedNoted) {
-        Bundle args = new Bundle();
-        args.putParcelable("note", Parcels.wrap(selectedNoted) );
-        NoteContentFragment contentFragment = new NoteContentFragment();
-        contentFragment.setArguments(args);
-
-//        fragmentManager
-//                .beginTransaction()
-//                .replace(R.id.fragmentContainer, contentFragment)
-//                .addToBackStack(null)
-//                .commit();
-    }
-
-    @Override
-    public void onBackPressed() {
-        if(getFragmentManager().getBackStackEntryCount() == 0) {
-            super.onBackPressed();
-        }
-        else {
-            getFragmentManager().popBackStack();
-        }
-    }
-
-
 
 }
