--- app/src/main/java/com/example/fbuteamproject/activities/MainActivity.java
@@ -1,43 +1,0 @@
-package com.example.fbuteamproject.activities;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.example.fbuteamproject.R;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-
-        Button button;
-
-        button = findViewById(R.id.button);
-        button.setOnClickListener(new View.OnClickListener() {
-
-
-            @Override
-            public void onClick(View v) {
-
-                Intent intent = new Intent(getApplicationContext(), PhotoSceneView.class);
-                startActivity(intent);
-            }
-
-            });
-
-
-    }
-
-
-
-
-
-}
-
-
