--- app/src/main/java/com/example/fbuteamproject/interfaces/NoteTitleSelectedInAdapterListener.java
@@ -1,6 +1,0 @@
-package com.example.fbuteamproject.interfaces;
-
-public interface NoteTitleSelectedInAdapterListener {
-
-    public void getNoteTitlePosition(int position);
-}
