--- app/src/main/java/com/example/fbuteamproject/interfaces/PassNoteToActivityListener.java
@@ -1,9 +1,0 @@
-package com.example.fbuteamproject.interfaces;
-
-import com.example.fbuteamproject.models.Note;
-
-public interface PassNoteToActivityListener {
-
-    public void returnSelectedNote(Note selectedNoted);
-
-}
