--- app/src/main/res/layout/activity_ar.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activities.ARActivity">
-
-    <com.google.ar.sceneform.ArSceneView
-        android:id="@+id/ar_scene_view"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:layout_gravity="top"/>
-
-</RelativeLayout>
-
-
-
-
-
-<!--<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"-->
-    <!--xmlns:app="http://schemas.android.com/apk/res-auto"-->
-    <!--xmlns:tools="http://schemas.android.com/tools"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="match_parent"-->
-    <!--tools:context=".activities.ARActivity">-->
-
-    <!--<FrameLayout-->
-        <!--android:id="@+id/fragmentContainer"-->
-        <!--android:layout_width="0dp"-->
-        <!--android:layout_height="0dp"-->
-        <!--android:layout_marginStart="4dp"-->
-        <!--android:layout_marginLeft="4dp"-->
-        <!--android:layout_marginTop="2dp"-->
-        <!--android:layout_marginEnd="4dp"-->
-        <!--android:layout_marginRight="4dp"-->
-        <!--android:layout_marginBottom="2dp"-->
-        <!--app:layout_constraintBottom_toBottomOf="parent"-->
-        <!--app:layout_constraintEnd_toEndOf="parent"-->
-        <!--app:layout_constraintStart_toStartOf="parent"-->
-        <!--app:layout_constraintTop_toTopOf="parent">-->
-
-    <!--</FrameLayout>-->
-<!--</android.support.constraint.ConstraintLayout>-->
