--- app/src/main/res/layout/activity_ux.xml
@@ -1,33 +1,0 @@
-<!--
-   Copyright 2018 Google LLC.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
--->
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activities.PhotosActivity">
-
-  <fragment android:name="com.google.ar.sceneform.ux.ArFragment"
-      android:id="@+id/ar_holder"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent" />
-
-  <!--<com.google.ar.sceneform.ArSceneView-->
-      <!--android:id="@+id/ux_fragment"-->
-      <!--android:layout_width="match_parent"-->
-      <!--android:layout_height="match_parent" />-->
-
-
-</FrameLayout>
