--- app/src/main/res/layout/player_view.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-<com.google.android.exoplayer2.ui.PlayerView
-    android:id="@+id/video_view"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    ></com.google.android.exoplayer2.ui.PlayerView>
-
-</LinearLayout>
